move molgraph featurizers over to generics
