+++ chemprop/featurizers/__init__.py
@@ -1,4 +1,5 @@
-from .atom import MultiHotAtomFeaturizer, AtomFeaturizer
-from .bond import MultiHotBondFeaturizer, BondFeaturizer
-from .molgraph import *
-from .molecule import *
+from .base import Featurizer, S, T
+from .atom import MultiHotAtomFeaturizer
+from .bond import MultiHotBondFeaturizer
+# from .molgraph import *
+# from .molecule import *
