+++ chemprop/featurizers/molgraph/__init__.py
@@ -1,5 +1,4 @@
-from .base import MolGraphFeaturizer
+# from .base import MolGraphFeaturizer
 from .cache import MolGraphCacheFacade, MolGraphCache, MolGraphCacheOnTheFly
-from .molgraph import MolGraph
 from .molecule import SimpleMoleculeMolGraphFeaturizer
 from .reaction import CondensedGraphOfReactionFeaturizer, CGRFeaturizer, RxnMode
